extract cell and notebook models
